adapted tests to new structure
